Avoid undefined behavior from signed integer shift wraparound
Signed integer wraparound is undefined behavior in according to
standards since at least C99, and that's what happens if you bit-shift
a 32bit integer with 31. And integer literals are signed by default,
which is one of those things you never think about it until you run
head-long into it.

In the old days, compilers were more about "do what I mean" than "look,
there's a loophole in the standard we can blame on the programmer and
call it optimization". Sigh.

Use an explicitly unsigned type where needed, but this may cause issues
elsewhere. Like here, the popt struct member whose type we have no
means to change is a signed integer. But we can't really just leave
undefined behavior in our public headers forever. So just in case this
does break something: thank the compiler committees.

Originally reported by Demi Marie Obenour in the form of a PR and now
again in OpenScanHub Fedora scan.

+++ lib/poptI.cc
@@ -129,7 +129,7 @@ struct poptOption rpmInstallPoptTable[] = {
 	N_("relocate files in non-relocatable package"), NULL},
 
  { "deploops", '\0', POPT_BIT_SET|POPT_ARGFLAG_DOC_HIDDEN,
- 	&rpmIArgs.transFlags, RPMTRANS_FLAG_DEPLOOPS,
+	&rpmIArgs.transFlags, (int)RPMTRANS_FLAG_DEPLOOPS,
 	N_("print dependency loops as warning"), NULL},
 
  { "erase", 'e', POPT_BIT_SET,
